+++ skippy/src/main/kotlin/com/slack/skippy/ComputeAffectedProjectsCli.kt
@@ -51,7 +51,7 @@ public class ComputeAffectedProjectsCli :
 
   private val mergeOutputs: Boolean by
     option("--merge-outputs", help = "Merge outputs from all configs into a single /merged dir.")
-      .flag(default = true)
+      .flag("--no-merge-outputs", default = true)
 
   private val config: Path by
     option(
@@ -63,7 +63,8 @@ public class ComputeAffectedProjectsCli :
       .required()
 
   private val computeInParallel: Boolean by
-    option("--parallel", help = "Compute affected projects in parallel.").flag(default = true)
+    option("--parallel", help = "Compute affected projects in parallel.")
+      .flag("--no-parallel", default = true)
 
   private val changedFiles: Path by
     option(